Avoid repetition of full classes, factoring out bits specific to specs2 version
