+++ core/src/main/scala-2/net/ruippeixotog/akka/testkit/specs2/impl/CompatImplicits.scala
@@ -1,0 +1,3 @@
+package net.ruippeixotog.akka.testkit.specs2.impl
+
+object CompatImplicits {}
