+++ core/src/main/scala-3/net/ruippeixotog/akka/testkit/specs2/impl/CompatMatchers.scala
@@ -1,0 +1,27 @@
+package net.ruippeixotog.akka.testkit.specs2.impl
+
+import net.ruippeixotog.akka.testkit.specs2.api.ReceiveMatcher
+import net.ruippeixotog.akka.testkit.specs2.impl.Matchers.TimeoutFunc
+import net.ruippeixotog.akka.testkit.specs2.impl.Matchers.GetMessageFunc
+import org.specs2.matcher.Expectable
+import org.specs2.execute.Result
+import org.specs2.matcher.ValueCheck
+import org.specs2.matcher.ExpectedResults.ok
+import org.specs2.matcher.ValueChecks.given
+import net.ruippeixotog.akka.testkit.specs2.Util._
+import org.specs2.matcher.Matchers._
+import org.specs2.execute.AsResult.given
+
+object CompatMatchers {
+  abstract class ReceiveMatcherImpl[P, A](implicit tf: TimeoutFunc[P]) extends ReceiveMatcher[P, A] {
+    def getMessage: GetMessageFunc[P, A]
+
+    def apply[S <: P](t: Expectable[S]): Result =
+      getMessage(t.value, tf(t.value)).result
+  }
+
+  def getRemainingMessages[P, A](_getMessage: GetMessageFunc[P, A], remMsgs: Seq[A]): GetMessageFunc[P, A] =
+    _getMessage.andThen(_.mapCheck(beOneOf[A](remMsgs *).check))
+
+  def partialToOk[A, B](f: PartialFunction[A, B]): ValueCheck[A] = f.andThen(_ => ok)
+}
