+++ core/src/main/scala/net/ruippeixotog/akka/testkit/specs2/impl/ClassicMatchers.scala
